Move cfp admin work to own blueprint
