+++ templates/admin/_nav.html
@@ -15,32 +15,34 @@
     <li class="dropdown">
         <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button"
             aria-haspopup="true" aria-expanded="false">Payments
-            {{badge(unreconciled_count + expiring_count)}}
+            {# {{badge(unreconciled_count + expiring_count)}} #}
             <span class="caret"></span></a>
         <ul class="dropdown-menu">
-            {{ menuitem("All", ".payments") }}
-            {{ menuitem("Bank Transfer", ".transactions", unreconciled_count) }}
-            {{ menuitem("Expiring", ".expiring", expiring_count) }}
-            {{ menuitem("Suppressed Transactions", ".suppressed") }}
+            {{ menuitem("All", "admin.payments") }}
+            {# {{ menuitem("Bank Transfer", "admin.transactions", unreconciled_count) }} #}
+            {{ menuitem("Bank Transfer", "admin.transactions") }}
+            {# {{ menuitem("Expiring", "admin.expiring", expiring_count) }} #}
+            {{ menuitem("Expiring", "admin.expiring") }}
+            {{ menuitem("Suppressed Transactions", "admin.suppressed") }}
         </ul>
     </li>
     <li class="dropdown">
         <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button"
             aria-haspopup="true" aria-expanded="false">Tickets <span class="caret"></span></a>
         <ul class="dropdown-menu">
-            {{ menuitem("Transfers", ".ticket_transfers") }}
-            {{ menuitem("Ticket Types", ".ticket_types") }}
+            {{ menuitem("Transfers", "admin.ticket_transfers") }}
+            {{ menuitem("Ticket Types", "admin.ticket_types") }}
         </ul>
     </li>
     <li class="dropdown">
         <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button"
             aria-haspopup="true" aria-expanded="false">CfP <span class="caret"></span></a>
         <ul class="dropdown-menu">
-            {{ menuitem("Proposals", ".cfp_proposals") }}
-            {{ menuitem("Categories", ".cfp_categories") }}
+            {{ menuitem("Proposals", "cfp_review.cfp_proposals") }}
+            {{ menuitem("Categories", "cfp_review.cfp_categories") }}
         </ul>
     </li>
-    {{ menuitem("Users", ".users") }}
-    {{ menuitem("Feature Flags", ".feature_flags") }}
+    {{ menuitem("Users", "admin.users") }}
+    {{ menuitem("Feature Flags", "admin.feature_flags") }}
 {% endif %}
 </ul>
